Moved half of the tests to src. This provides better flow
