+++ homepairs/homepairsUI/babel.config.js
@@ -30,12 +30,13 @@ module.exports = function(api) {
                         'homepairs-elements': './src/Elements/index.tsx',
                         'homepairs-redux-actions': './src/state/actions.ts',
                         'homepairs-pages': './src/Screens/index.tsx',
-                        'homepairs-modals': './src/Modals/index',
+                        'homepairs-modals': './src/Modals/index.tsx',
                         'homepairs-base-styles': './res/Styles/base.ts',
-                        'homepairs-components': './src/Components/index.ts',
+                        'homepairs-components': './src/Components/index.tsx',
                         'homepairs-utilities': './src/utility/index.tsx',
                         'homepairs-routes' : './src/Routes/RouteConstants',
                         'homepairs-endpoints' : './src/Routes/RemoteEndpoints.tsx',
+                        'homepairs-test' : './tests/fixtures/index.tsx',
                     },
                 },
             ],
