+++ homepairs/homepairsUI/src/Components/HomePairsHeader/HomePairsNavHeaderTest/HomePairsHeaderDropDown.test.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
-import { propertyManagerMock1, HeaderMockStores } from 'tests/fixtures/StoreFixture';
-import { HomePairsHeader } from 'homepairs-components';
-import { prepareNavigationStackFirstRouteMock, prepareNavigationMock } from 'tests/fixtures/DummyComponents';
+import { 
+    propertyManagerMock1, 
+    HeaderMockStores,
+    prepareNavigationStackFirstRouteMock, 
+    prepareNavigationMock } from 'homepairs-test';
 import { fireEvent, render } from 'react-native-testing-library';
 import { Provider } from 'react-redux';
-import { HEADER_ACTION_TYPES } from 'src/state/header/actions';
+import { HEADER_ACTION_TYPES } from 'homepairs-redux-actions';
 import { 
     HeaderAction, 
     UpdateSelectedPageAction, 
@@ -13,8 +15,9 @@ import {
     MainAppStackType,
 } from 'homepairs-types';
 import { TouchableOpacity, Platform} from 'react-native';
-import { navigationPages } from 'src/Routes/RouteConstants';
-import {BrowserRouter as Router} from 'react-router-dom';
+import { BrowserRouter as Router } from 'react-router-dom';
+import { navigationPages } from '../../../Routes/RouteConstants';
+import HomePairsHeader from '../HomePairsHeader';
 
 
 const [mockStackNavigation, navigationStackSpyFunction] = prepareNavigationMock();
