+++ homepairs/homepairsUI/src/state/actions.ts
@@ -5,5 +5,13 @@ import * as ServiceActions from './service/actions';
 import * as SessionActions from './session/actions';
 import * as PreferredProviderActions from './preferred-service-provider/actions';
 
+
 export {AccountActions, HeaderActions, PropertyListActions, ServiceActions, SessionActions, PreferredProviderActions} ;
 export const HomePairsStateActions = {AccountActions, HeaderActions, PropertyListActions, ServiceActions};
+
+export * from './account/actions';
+export * from './header/actions';
+export * from './property-list/actions';
+export * from './service/actions';
+export * from './session/actions';
+export * from './preferred-service-provider/actions';
