+++ homepairs/homepairsUI/tests/fixtures/index.tsx
@@ -1,0 +1,3 @@
+export * from './DummyComponents';
+export * from './StoreFixture';
+export {default as StyleFixture} from './StyleFixture';
