+++ homepairs/homepairsUI/tsconfig.json
@@ -24,15 +24,15 @@
             "homepairs-elements": ["./src/Elements/index.tsx"],
             "homepairs-redux-actions": ["./src/state/actions.ts"],
             "homepairs-pages": ["src/Screens/index.tsx"],
-            "homepairs-modals": ["src/Modals/index"],
+            "homepairs-modals": ["src/Modals/index.tsx"],
             "homepairs-base-styles": ["./res/Styles/base.ts"],
             "homepairs-utilities": ["./src/utility/index.tsx"],
-            "homepairs-components": ["src/Components/index.ts"],
+            "homepairs-components": ["src/Components/index.tsx"],
             "homepairs-routes" : ["./src/Routes/RouteConstants"],
             "homepairs-endpoints" : ["./src/Routes/RemoteEndpoints.tsx"],
-
+            "homepairs-test" : ["./tests/fixtures/index.tsx"],
         }
     },
     "exclude": ["node_modules", "../tests", "build"],
-    "include": ["src", "res", "tests/redux/property-list/GenerateTenantAccount.test.tsx", "tests/pages/Components/SceneHeader.test.tsx", "tests/pages/Components/HomePairsNavHeader/HomePairsHeaderDropDown.test.tsx"]
+    "include": ["src", "res"]src/Components/HomePairsHeader/HomePairsNavHeader/HomePairsHeasrc/Components/SceneHeader/SceneHeader.test.tsx
 }
